# api/OpenAPI.json
{
    "info": {
        "title": "modifiability-scenario-creator-api",
        "version": "0.0.2",
        "license": {
            "name": "Apache-2.0"
        },
        "description": "[Specification JSON](/api-spec)\n\nAPI to create and evaluate software modifiability scenarios"
    },
    "paths": {
...
        "/changes/{id}": {
            "get": {
                "summary": "/changes/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            },
            "put": {
                "summary": "/changes/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "/changes/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            }
        },
...
        "/scenarios/{id}": {
            "get": {
                "summary": "/scenarios/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            },
            "put": {
                "summary": "/scenarios/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "/scenarios/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            }
        },
        "/scenario/{id}": {
            "get": {
                "summary": "/scenario/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            }
        },
...
        "/services/{id}": {
            "get": {
                "summary": "/services/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            },
            "put": {
                "summary": "/services/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "summary": "/services/{id}",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "100": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "swagger": "2.0"
}